+++ src/baobab_scryfall_api_caller/models/bulk_data/__init__.py
@@ -1,0 +1,5 @@
+"""Modeles du domaine Bulk Data."""
+
+from baobab_scryfall_api_caller.models.bulk_data.bulk_data import BulkData
+
+__all__ = ["BulkData"]